# Conflicts:
#	src/CoreBundle/Settings/SessionSettingsSchema.php

+++ public/main/admin/course_list.php
@@ -12,6 +12,7 @@ declare(strict_types=1);
 use Chamilo\CoreBundle\Component\Utils\ActionIcon;
 use Chamilo\CoreBundle\Component\Utils\StateIcon;
 use Chamilo\CoreBundle\Component\Utils\ToolIcon;
+use Chamilo\CoreBundle\Framework\Container;
 
 $cidReset = true;
 
@@ -244,7 +245,11 @@ function get_course_data(
                 ICON_SIZE_SMALL,
                 get_lang('Delete')
             ),
-            $path.'admin/course_list.php?delete_course='.$course['col0'],
+            $path.'admin/course_list.php?'
+            .http_build_query([
+                'delete_course' => $course['col0'],
+                'sec_token' => Security::getTokenFromSession(),
+            ]),
             [
                 'onclick' => "javascript: if (!confirm('"
                     .addslashes(api_htmlentities(get_lang('Please confirm your choice'), \ENT_QUOTES))
@@ -319,7 +324,7 @@ function get_course_visibility_icon(int $visibility): string
     };
 }
 
-if (isset($_POST['action'])) {
+if (isset($_POST['action']) && Security::check_token('get')) {
     // Delete selected courses
     if ('delete_courses' == $_POST['action']) {
         if (!empty($_POST['course'])) {
@@ -392,7 +397,7 @@ if (isset($_GET['search']) && 'advanced' === $_GET['search']) {
     $content .= $form->returnForm();
 } else {
     $tool_name = get_lang('Course list');
-    if (isset($_GET['delete_course'])) {
+    if (isset($_GET['delete_course']) && Security::check_token('get')) {
         $result = CourseManager::delete_course($_GET['delete_course']);
         if ($result) {
             Display::addFlash(Display::return_message(get_lang('Deleted')));
@@ -416,9 +421,12 @@ if (isset($_GET['search']) && 'advanced' === $_GET['search']) {
         ['id' => 'course-search-keyword', 'aria-label' => get_lang('Search courses')]
     );
     $form->addButtonSearch(get_lang('Search courses'));
-    $advanced = '<a class="btn btn--plain" href="'.api_get_path(WEB_CODE_PATH).'admin/course_list.php?search=advanced">
-        <em class="pi pi-search"></em> '.
-        get_lang('Advanced search').'</a>';
+    $advanced = Display::toolbarButton(
+        get_lang('Advanced search'),
+        Container::getRouter()->generate('legacy_main', ['name' => 'admin/course_list.php', 'search' => 'advanced']),
+        ActionIcon::SEARCH,
+        'plain'
+    );
 
     // Create a filter by session
     $sessionFilter = new FormValidator(
@@ -500,6 +508,7 @@ if (isset($_GET['search']) && 'advanced' === $_GET['search']) {
     );
 
     $parameters = [];
+    $parameters['sec_token'] = Security::get_token();
     if (isset($_GET['keyword'])) {
         $parameters = ['keyword' => Security::remove_XSS($_GET['keyword'])];
     } elseif (isset($_GET['keyword_code'])) {
